Finalize implementation of the use of Gauss-Lobatto nodes for the incompressible flow solver with SDC time integration.
The problem with Gauss-Lobatto nodes is the fact that the first evaluation is explicit instead of implicit.
The velocity on the faces of each cell cannot be evaluated directly in OpenFOAM. Therefore, the best approach
would be to startup a computation with Gauss-Radau points, for which each evalauation is implicit.
The evaluation of UfF and UF at the last time step is equal to the first time step, so this is stored on disk.
A computation with Gauss-Lobatto nodes is then restarted from this result, and has the function evaluation on the faces
of the cell available.

Closes #45.