Add first-participant setting to fsi.yaml.
Possible options: fluid-solver or solid-solver
The first participant is the solver which is executed first in case
of a staggered coupling procedure. In case the solid solver is executed
first, the convergence measure is based upon the traction acting on the
FSI interface. When the fluid solver is the first participant, the
FSI residual is based on the displacement of the FSI interface.